requirements.txt
flask
waitress
torch
Pillow
transformers
flask-cors